Solved issue where generating random graph broke
Wasn't working because verties would spawn on the same location and they
wouldn't have any distance inbetween them. I fixed it so that they
always have decimal points and therefore distance inbetween. There is
still a tiny change that they have the same location but the chances of
that happening are very tiny.